Implements hdual class in HDual file
This is an extension of the dual numbers useful for calculating higher
order derivatives. Has all the same methods as dual. As of right now
arithmetic can only be done between dual numbers of the same dimension
and between dual numbers and real numbers. Future updates should allow
for arithmetic between dual numbers of unequal dimension. Also adds
is_real and conj methods to both dual and hdual objects, which check if
the imaginary parts are zero and calculates the conjugate respectively.