Updates exp, sin, and cos functions for hduals.
Exponential function is highly inefficient. Consider optimizing for
future versions. Added functionality to use exponential with complex
numbers. As of right now, this doesn't work with regular dual numbers.
This is because it is intended only to help calculate sine and cosine.
Consider changing in future versions if there is demand for it. Changed
hdual object so that it can now handle complex numbers.