Implements EDiffeqs file.
EDiffeqs is a file for solving ordinary differential equations. It
includes a tseries object which is an object that either calculates a
Taylor Series for a given analytic function or builds a Taylor Series
from a list of given coefficients. If called, it produces the value of
the Taylor Series for a given input.

Also has sseries function for producing series solutions for linear
ordinary differential equations. This method can likely be generalized
for certain non-linear ODEs as well, further investigations will need to
be done to confirm this.

Fixed error with hdual object that caused negation to behave
incorrectly.

Fixed error with sin and cos function that caused them to break when
given an hdual with only a single input.